Critic problems correction and big hotfix

+++ src/app/actions/deleteCatalogs.js
@@ -3,15 +3,24 @@
 import { db } from "../utils/firebase";
 import { deleteDoc, collection, doc, getDocs, query, where } from "firebase/firestore";
 import { deleteWhatsappSession } from "./deleteWhatsappSession";
+import { deleteStorageDirectory } from "./deleteStorageDirectory";
 
-export async function deleteCatalogs(catalogs) {
+export async function deleteCatalogs(catalogs, userId) {
+    console.log(catalogs)
     for (const catalog of catalogs) {
-        const q = query(collection(db, "products"), where("catalog", "==", catalog.id));
+        const catalogId = catalog.id
+        
+        await deleteStorageDirectory(`${userId}/${catalogId}`);
+
+        const q = query(collection(db, "products"), where("catalog", "==", catalogId));
         const querySnapshot = await getDocs(q);
-        for (const doc of querySnapshot.docs) {
-            await deleteDoc(doc.ref);
+        if (!querySnapshot.empty) {
+            for (const doc of querySnapshot.docs) {
+                await deleteDoc(doc.ref);
+            }
         }
-        await deleteWhatsappSession(catalog.whatsapp_session);
-        await deleteDoc(doc(db, "catalogs", catalog.id));
+        if(catalog.whatsapp_session) await deleteWhatsappSession(catalog.whatsapp_session);
+        await deleteDoc(doc(db, "catalogs", catalogId));
+        
     }
 }
